fix: resolve all compilation and test errors
- Add NetworkModule (Gson, OkHttpClient, Retrofit, OpenAiApi, modelName)
- Fix ScreenCaptureProvider: use getBoundsInScreen(Rect) instead of bounds property
- Fix MemoEditScreen: replace non-existent TimePickerDialog with AlertDialog+TimePicker
- Fix OpenAiLlmClient: add @Named("modelName") qualifier
- Fix AudioRecorder: Int to Short cast in writeShort
- Fix VoiceMemoService: add missing withTimeoutOrNull import
- Fix VolumeKeyAccessibilityService: add Rect import and getBoundsInScreen
- Fix MemoUseCasesTest: rewrite with MockK instead of Mockito, remove title param
- Build passes, all unit tests pass

+++ app/src/main/java/com/qb/secondbrain/service/AudioRecorder.kt
@@ -134,7 +134,7 @@ class AudioRecorder(
             raf.write(shortToLittleEndian(1))             // channels: mono
             raf.write(intToLittleEndian(sampleRate))      // sample rate
             raf.write(intToLittleEndian(sampleRate * BYTES_PER_SAMPLE)) // byte rate
-            raf.write(shortToLittleEndian(BYTES_PER_SAMPLE))            // block align
+            raf.write(shortToLittleEndian(BYTES_PER_SAMPLE.toShort()))  // block align
             raf.write(shortToLittleEndian(16))            // bits per sample
 
             // data sub-chunk
